15/11

+++ src/main/java/vn/com/devmaster/service/managermaterial/domain/OrdersTransport.java
@@ -25,7 +25,7 @@ public class OrdersTransport {
     private TransportMethod idtransport;
 
     @Column(name = "TOTAL")
-    private Integer total;
+    private Double total;
 
     @Column(name = "NOTES")
     private Integer notes;
